feature: [BE] running 서버 swift, kotlin 지원
