+++ backEnd/running/src/common/type.ts
@@ -1,2 +1,1 @@
-export type supportLang = 'python' | 'javascript' | 'java' | 'c';
 export type runCommandResult = { stdout: string; stderr: string };
