+++ backEnd/running/src/common/utils.ts
@@ -21,18 +21,3 @@ export const Messages = {
   TIMEOUT: '코드가 실행되는데 너무 오래 걸립니다.',
   UNKNOWN: '알 수 없는 에러가 발생했습니다.',
 };
-
-export function languageCommand(language, filePaths) {
-  const [filepath, compile_dist] = filePaths;
-  switch (language) {
-    case 'python':
-      return `python3 ${filepath}`;
-    case 'javascript':
-      return `node ${filepath}`;
-    case 'java':
-      return `java ${filepath}`;
-    case 'c':
-      return `gcc -o ${compile_dist} ${filepath} && ${compile_dist}`;
-  }
-}
-export const needCompile = ['c'];
